Move Network parser code into Network impl
